LayoutTests/imported/w3c/web-platform-tests/css/css-anchor-position/last-successful-change-try-rule-expected.txt


PASS Starts rendering with --try
PASS No successful position, keep --try
FAIL No successful position, last successful invalidated by @position-try change assert_equals: expected 0 but got 200
